requirements.txt
aniso8601==8.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysqlclient==2.0.3
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.3
Werkzeug==1.0.1
